ISSUE #327 * Move classes to a different module hierarchy.
