+++ minos/common/model/__init__.py
@@ -6,6 +6,31 @@ This file is part of minos framework.
 Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
 """
 from .abc import (
+    MinosModel,
+)
+from .aggregate import (
+    Aggregate,
+)
+from .command import (
+    Command,
+)
+from .command_reply import (
+    CommandReply,
+)
+from .dynamic import (
+    DynamicMinosModel,
+)
+from .event import (
+    Event,
+)
+from .fields import (
+    MinosModelAvroDataBuilder,
+    MinosModelAvroSchemaBuilder,
+    MinosModelFromAvroBuilder,
+    ModelField,
+    ModelFieldCaster,
+)
+from .types import (
     ARRAY,
     BOOLEAN,
     BYTES,
@@ -32,23 +57,6 @@ from .abc import (
     Decimal,
     Enum,
     Fixed,
-    MinosModel,
     MissingSentinel,
-    ModelField,
     ModelRef,
-)
-from .aggregate import (
-    Aggregate,
-)
-from .command import (
-    Command,
-)
-from .command_reply import (
-    CommandReply,
-)
-from .dynamic import (
-    DynamicMinosModel,
-)
-from .event import (
-    Event,
 )
