--- minos/common/model/abc/__init__.py
@@ -1,47 +1,0 @@
-"""
-Copyright (C) 2021 Clariteia SL
-
-This file is part of minos framework.
-
-Minos framework can not be copied and/or distributed without the express permission of Clariteia SL.
-"""
-from .fields import (
-    MinosModelAvroDataBuilder,
-    MinosModelAvroSchemaBuilder,
-    MinosModelFromAvroBuilder,
-    ModelField,
-    ModelFieldCaster,
-)
-from .model import (
-    MinosModel,
-)
-from .types import (
-    ARRAY,
-    BOOLEAN,
-    BYTES,
-    CUSTOM_TYPES,
-    DATE,
-    DECIMAL,
-    DOUBLE,
-    ENUM,
-    FIXED,
-    FLOAT,
-    INT,
-    LONG,
-    MAP,
-    NULL,
-    PYTHON_ARRAY_TYPES,
-    PYTHON_IMMUTABLE_TYPES,
-    PYTHON_LIST_TYPES,
-    PYTHON_NULL_TYPE,
-    PYTHON_TYPE_TO_AVRO,
-    STRING,
-    TIME_MILLIS,
-    TIMESTAMP_MILLIS,
-    UUID,
-    Decimal,
-    Enum,
-    Fixed,
-    MissingSentinel,
-    ModelRef,
-)
