+++ minos/common/model/fields/utils.py
@@ -10,7 +10,7 @@ import typing as t
 
 
 def _is_minos_model_cls(type_field: t.Type) -> bool:
-    from ..model import (
+    from ..abc import (
         MinosModel,
     )
 
@@ -18,7 +18,7 @@ def _is_minos_model_cls(type_field: t.Type) -> bool:
 
 
 def _is_aggregate_cls(type_field: t.Type) -> bool:
-    from ...aggregate import (
+    from ..aggregate import (
         Aggregate,
     )
